requirements.txt
awsebcli==3.7.6
boto==2.40.0
botocore==1.4.27
cement==2.8.2
colorama==0.3.7
Django==1.8.3
django-appconf==1.0.2
django-common-helpers==0.8.0
django-cron==0.4.6
django-imagekit==3.3
django-paypal==0.3.2
django-social-auth==0.7.28
django-storages==1.4.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
httplib2==0.9.2
jmespath==0.9.0
oauth2==1.9.0.post1
oauthlib==1.1.2
pathspec==0.3.4
pilkit==1.1.13
Pillow==3.2.0
psycopg2==2.6.1
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2016.4
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
semantic-version==2.5.0
six==1.10.0
stripe==1.35.0
texttable==0.8.4
tweepy==3.5.0
websocket-client==0.37.0
wheel==0.24.0
